feat(#20): 전역 Provider 컴포넌트 생성
