+++ src/components/Provider/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Created on Fri Nov 24 2023
+ *
+ * Copyright (c) 2023 Your Company
+ */
+
+export { default as DonguramiProvider } from "./DonguramiProvider";
