+++ src/pages/_app.tsx
@@ -1,21 +1,18 @@
-import colorTheme from "@/styles/color";
-import fontTheme from "@/styles/font";
-import "@/styles/globals.css";
-import { Theme, ThemeProvider } from "@emotion/react";
-import type { AppProps } from "next/app";
-import { RecoilRoot } from "recoil";
+/*
+ * Created on Fri Nov 24 2023
+ *
+ * Copyright (c) 2023 Your Company
+ */
 
-const theme: Theme = {
-  color: colorTheme,
-  font: fontTheme,
-};
+import type { AppProps } from "next/app";
+
+import "@/styles/globals.css";
+import { DonguramiProvider } from "@/components";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
-    <RecoilRoot>
-      <ThemeProvider theme={theme}>
-        <Component {...pageProps} />
-      </ThemeProvider>
-    </RecoilRoot>
+    <DonguramiProvider>
+      <Component {...pageProps} />
+    </DonguramiProvider>
   );
 }
